refactor: Modularize auth controller and fix cross-domain cookie issues
 Major Refactoring:
- Extract cookie configuration to utils/cookieConfig.js
- Extract GitHub OAuth logic to services/githubOAuth.js
- Extract user management to services/userService.js
- Simplify auth controller from ~600 to ~150 lines

 Cross-Domain Cookie Fixes:
- Set domain: undefined for cross-origin scenarios
- Proper SameSite=none configuration for production
- Enhanced cookie logging for debugging

 Production Cleanup:
- Remove debug endpoints and test components
- Remove unused debug utilities
- Clean up frontend dashboard component

 New Files:
- backend/src/utils/cookieConfig.js - Centralized cookie config
- backend/src/services/githubOAuth.js - GitHub API interactions
- backend/src/services/userService.js - User management logic
- REFACTORING_SUMMARY.md - Complete refactoring documentation
- test-auth-locally.md - Local testing instructions

 Completed Tasks:
- Task 9: Fix cross-domain cookie transmission issue
- Task 10: Refactor auth controller for production readiness

 Ready for Production:
- Modular, maintainable codebase
- Proper cross-domain authentication
- No debug code in production build
